fix phpstan errors
