+++ tests/PHPStan/Node/data/attributes.php
@@ -1,4 +1,4 @@
-<?php
+<?php // lint >= 8.0
 
 namespace NodeCallbackCalled;
